# Optimize day 6