--- a/santa/src/Santa/Runner.hs
+++ b/santa/src/Santa/Runner.hs
@@ -2,17 +2,18 @@ module Santa.Runner
   ( benchAll,
     fmt,
     run,
-  ) where
+  )
+where
 
-import Criterion.Types (Config(..))
 import Criterion.Main
+import Criterion.Types (Config (..))
 import qualified Data.Map as Map
 import qualified Data.Text as Text
 import Protolude
-import Text.Printf (printf)
 import Santa.Problems (Problems)
 import Santa.Solution (Error)
 import System.FilePath.Posix ((</>))
+import Text.Printf (printf)
 
 benchAll :: Problems -> FilePath -> IO ()
 benchAll problems inputs = do
@@ -25,11 +26,12 @@ benchAll problems inputs = do
       contents <- input inputs n
       return $ bench (Text.unpack n) $ nf solution contents
 
-    config = defaultConfig
-      { reportFile = Just "./bench/report.html"
-      , csvFile    = Just "./bench/report.csv"
-      , jsonFile   = Just "./bench/report.json"
-      }
+    config =
+      defaultConfig
+        { reportFile = Just "./bench/report.html",
+          csvFile = Just "./bench/report.csv",
+          jsonFile = Just "./bench/report.json"
+        }
 
 run :: Problems -> FilePath -> Text -> IO (Either Error (Text, Text))
 run problems inputs n = case Map.lookup n problems of