--- a/santa/src/Santa/Solution.hs
+++ b/santa/src/Santa/Solution.hs
@@ -1,6 +1,8 @@
 module Santa.Solution
   ( Solution,
     Error,
+    parse,
+    parse',
     part1,
     part2,
     run,
@@ -10,15 +12,17 @@ where
 
 import Control.Lens
 import Data.Attoparsec.Text (Parser, endOfInput, parseOnly)
-import Data.Text (pack)
+import qualified Data.Attoparsec.Text as Atto
 import Data.Either.Combinators (mapLeft)
+import Data.Text (pack)
 import Protolude
 
 type S = (Text, Text)
 
 newtype SolutionT m a = SolutionT
   { unSolutionT :: StateT S m a
-  } deriving (Functor, Applicative, Monad, MonadState S)
+  }
+  deriving (Functor, Applicative, Monad, MonadState S)
 
 data Solution' m a = Solution
   { parser :: Parser a,
@@ -31,13 +35,24 @@ newtype Error = ParseError Text
   deriving (Show, Eq, NFData)
 
 run :: Solution a -> Text -> Either Error S
-run solution input = runIdentity $ either
-  (return . Left . ParseError)
-  (\parsed -> Right <$> execStateT (unSolutionT $ runner solution parsed) ("", ""))
-  $ parse (parser solution) input
+run solution input =
+  runIdentity $
+    either
+      (return . Left . ParseError)
+      (\parsed -> Right <$> execStateT (unSolutionT $ runner solution parsed) ("", ""))
+      $ parse (parser solution) input
 
 parse :: Parser a -> Text -> Either Text a
-parse p i = mapLeft pack $ parseOnly (p <* optional "\n" <* endOfInput) i
+parse p = mapLeft pack . parseOnly (p <* optional "\n" <* endOfInput)
+
+-- parse' fails with a full description of the unconsumed input
+parse' :: Parser a -> Text -> Either Text a
+parse' p = fmt . Atto.parse (p <* optional "\n")
+  where
+    fmt (Atto.Done "" result) = Right result
+    fmt (Atto.Done rest _) = Left $ "Unconsumed input: " <> rest
+    fmt (Atto.Fail _ _ msg) = Left $ pack msg
+    fmt (Atto.Partial cc) = fmt $ cc ""
 
 solve :: Functor m => Parser a -> (a -> SolutionT m b) -> Solution' m a
 solve p r = Solution p (void . r)