Local ∆ of 5.710 ms => 507.2 μs

This is a _bit_ of a cheat because the new parser now _requires_ a terminating \n,
but not having to backtrack is a huge timesaver.

Cleans up a few parsing odds-and-ends along the way:

* the `parse` in tests now uses the underlying `Santa.Solution.parse`
* introduces a `parse'` which prints the unconsumed input (currently unused, but helpful in dev)

# Before

    time                 5.710 ms   (5.581 ms .. 5.834 ms)
                         0.993 R²   (0.985 R² .. 0.998 R²)
    mean                 5.845 ms   (5.751 ms .. 6.016 ms)
    std dev              389.4 μs   (221.3 μs .. 571.4 μs)

    total alloc =  11,312,656 bytes  (excludes profiling overheads)

    COST CENTRE    MODULE                         SRC                                                 %time %alloc

    >>=.\          Data.Attoparsec.Internal.Types Data/Attoparsec/Internal/Types.hs:(148,9)-(149,44)   25.0   17.8
    *>             Data.Attoparsec.Internal.Types Data/Attoparsec/Internal/Types.hs:189:5-26           16.7    3.4
    readTextDevice Data.Text.Internal.IO          libraries/text/Data/Text/Internal/IO.hs:133:39-64     8.3    0.0
    fmap.\.succ'   Data.Attoparsec.Internal.Types Data/Attoparsec/Internal/Types.hs:173:11-58           8.3   28.7
    >>=.\.succ'    Data.Attoparsec.Internal.Types Data/Attoparsec/Internal/Types.hs:148:13-76           8.3    7.3
    >>=            Data.Attoparsec.Internal.Types Data/Attoparsec/Internal/Types.hs:(147,5)-(149,44)    8.3    0.0
    run            Santa.Runner                   src/Santa/Runner.hs:(35,1)-(37,50)                    8.3    0.9
    anyAnswered    P06                            src/P06.hs:20:1-35                                    8.3    6.8
    allAnswered    P06                            src/P06.hs:23:1-63                                    8.3    4.6
    lengthOf       Data.Attoparsec.Text.Internal  Data/Attoparsec/Text/Internal.hs:546:1-27             0.0    2.6
    fromString     Data.Attoparsec.Text.Internal  Data/Attoparsec/Text/Internal.hs:92:5-32              0.0    5.7
    plus.\         Data.Attoparsec.Internal.Types Data/Attoparsec/Internal/Types.hs:(163,3)-(164,38)    0.0    6.4
    many.some_v    Data.Attoparsec.Internal.Types Data/Attoparsec/Internal/Types.hs:214:9-37            0.0    2.3
    fmap.\         Data.Attoparsec.Internal.Types Data/Attoparsec/Internal/Types.hs:(173,7)-(174,42)    0.0    6.2
    -              Data.Attoparsec.Internal.Types Data/Attoparsec/Internal/Types.hs:47:38-40            0.0    2.6
    +              Data.Attoparsec.Internal.Types Data/Attoparsec/Internal/Types.hs:47:38-40            0.0    2.6

# After

    time                 507.2 μs   (483.9 μs .. 536.5 μs)
                         0.977 R²   (0.959 R² .. 0.992 R²)
    mean                 537.7 μs   (512.0 μs .. 581.7 μs)
    std dev              114.8 μs   (79.15 μs .. 179.5 μs)

    total alloc =  11,312,632 bytes  (excludes profiling overheads)

    COST CENTRE    MODULE                         SRC                                                 %time %alloc

    CAF            Santa.Main                     <entire-module>                                      18.5    0.0
    readTextDevice Data.Text.Internal.IO          libraries/text/Data/Text/Internal/IO.hs:133:39-64    14.8    0.0
    fmap.\.succ'   Data.Attoparsec.Internal.Types Data/Attoparsec/Internal/Types.hs:173:11-58          14.8   28.7
    >>=.\.succ'    Data.Attoparsec.Internal.Types Data/Attoparsec/Internal/Types.hs:148:13-76          14.8    7.3
    hPutStrLn.\    Protolude.Show                 src/Protolude/Show.hs:46:21-42                        7.4    0.1
    CAF            GHC.IO.Encoding                <entire-module>                                       3.7    0.0
    fromString     Data.Attoparsec.Text.Internal  Data/Attoparsec/Text/Internal.hs:92:5-32              3.7    5.7
    plus.\         Data.Attoparsec.Internal.Types Data/Attoparsec/Internal/Types.hs:(163,3)-(164,38)    3.7    6.4
    >>=.\          Data.Attoparsec.Internal.Types Data/Attoparsec/Internal/Types.hs:(148,9)-(149,44)    3.7   17.8
    >>=            Data.Attoparsec.Internal.Types Data/Attoparsec/Internal/Types.hs:(147,5)-(149,44)    3.7    0.0
    +              Data.Attoparsec.Internal.Types Data/Attoparsec/Internal/Types.hs:47:38-40            3.7    2.6
    >>=.\.\        Options.Applicative.Types      src/Options/Applicative/Types.hs:272:46-63            3.7    0.0
    run            Santa.Runner                   src/Santa/Runner.hs:(35,1)-(37,50)                    3.7    0.9
    lengthOf       Data.Attoparsec.Text.Internal  Data/Attoparsec/Text/Internal.hs:546:1-27             0.0    2.6
    many.some_v    Data.Attoparsec.Internal.Types Data/Attoparsec/Internal/Types.hs:214:9-37            0.0    2.3
    fmap.\         Data.Attoparsec.Internal.Types Data/Attoparsec/Internal/Types.hs:(173,7)-(174,42)    0.0    6.2
    -              Data.Attoparsec.Internal.Types Data/Attoparsec/Internal/Types.hs:47:38-40            0.0    2.6
    *>             Data.Attoparsec.Internal.Types Data/Attoparsec/Internal/Types.hs:189:5-26            0.0    3.4
    anyAnswered    P06                            src/P06.hs:20:1-35                                    0.0    6.8
    allAnswered    P06                            src/P06.hs:23:1-63                                    0.0    4.6